feature/PBAO-243-Move-all-encryption-infrastructure-classes-into-EasyEncryption
- Fixing tests
